ligprep and dock adjusted for Open Source
Ligprep no longer uses Schrodinger functions and has been converted to
openbabel for the most part. Dock now will setup a docking file where
each line is a gnina command line for docking the ligand to the receptor
in the box defined by the crystal ligand. Created a slurm option to
create a tarfile and adjust the docking file to allow easy docking on
a slurm cluster.