Fix score padding
In the case of a ligand not having `max_poses` docked conformations we
would like to ensure that those conformations are never picked.
Therefore, the docking score (logit(CNNscore)) is padded with `np.nan`
and the energy term for the pairs of poses is padded with `-np.inf`.
Previously the docking score was padded with `np.inf` (causing missing
conformations to always have the max value) and the energy term was
padded with 0 (which may be a higher energy than pairs that have both
conformations available). The selection now uses `np.nanargmax` so
`np.nan` values are never selected.

This will prevent the missing conformations, denoted by the elements
with padding, from being selected by the EM algorithm.