Retrieve small molecule ligands from PDB
If the ligand is a small molecule with a valid 3 letter code then we
will retrieve the SDF directly from the PDB. Otherwise, we will extract
the ligand from the PDB file. If the ligand does not have a 3 letter
residue name, then it is likely a peptide ligand so a pdb file is fine.