Download ligand from PDB and bug fixes
Now we download the ligand directly from the PDB to ensure we have the
correct bond orders. If the ligand does not have a PDB ligand ID then we
follow the same procedure as normal.

Fixed issue with peptide ligands. Removed the peptide ligand completely
from the protein only structure.