Fix memory/hanging issue for MCSS multiprocessing
Now mcss of all pairs is calculated first using a reduced number of
processes. The mcss is stored in a global dictionary where the keys are
the SMARTS strings of the molecules in the pair. Then RMSD is calculated
in using all of the specified processes, where we have a new thread for
each unique ligand,ligand pair.